% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_turning_point.R
\name{estimate_turning_point_sd}
\alias{estimate_turning_point_sd}
\title{Estimate Turning Point Based on Proximity to Minimum or Maximum Outcome}
\usage{
estimate_turning_point_sd(
  data,
  outcome,
  x_var,
  group_var,
  direction = c("min", "max"),
  sd_multiplier = 1
)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{outcome}{Character string. Name of the y-variable.}

\item{x_var}{Character string. Name of the x-variable.}

\item{group_var}{Character string. Name of the grouping variable.}

\item{direction}{Character string. Either \code{"min"} or \code{"max"}.
If \code{"min"}, the function selects the maximum x where y is close to min(y).
If \code{"max"}, the function selects the maximum x where y is close to max(y).}

\item{sd_multiplier}{Numeric. Number of standard deviations used to define
the cutoff. Default is 1.}
}
\value{
A data.table with one row per group and the selected turning_point.
}
\description{
For each group, this function finds the maximum x-value for which the
corresponding y-value is within \code{sd_multiplier} standard deviations of either
the minimum or maximum y-value.
}
