% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexDesignWindows-package.R
\docType{package}
\name{fcipRateRevisions-package}
\alias{fcipRateRevisions}
\alias{fcipRateRevisions-package}
\title{fcipRateRevisions: indexDesignWindows}
\description{
Replication package for studies examining the impacts of rating revisions in United States agricultural insurance. - Actuarial Impacts of Rating Revisions in United States Agricultural Insurance. - Actuarial Updates and Taxpayer Costs in Subsidized Agricultural Insurance Programs. - The Effect of Actuarial Updates on Private-to-Public Risk Transfers in Agricultural Insurance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ftsiboe/indexDesignWindows}
  \item Report bugs at \url{https://github.com/ftsiboe/indexDesignWindows/issues}
}

}
\keyword{internal}
